requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
Django==3.2.7
django-appconf==1.0.4
django-axes==5.25.0
django-crontab==0.7.1
django-ipware==4.0.0
djangorestframework==3.12.4
idna==3.2
importlib-metadata==4.8.1
jsonschema==3.2.0
mysqlclient==2.1.0
pyasn1==0.4.8
pycryptodome==3.10.1
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.6
zipp==3.5.0
uWSGI==2.0.20